Function processing keywords found in user input
The function is applied per keyword (which will be passed from the identify-keyword function) that has been found in the user input. This keyword is searched in the map containing information about the park (here I simply use a map directly, named "park-map" for convenience, will be changed to the parsed JSON file later)
Depending on the value this keyword has in the map - "yes"/"no" or the value itself in case of "attractions" and "transportation" fields - a corresponding sentence is printed.
There are several layouts of sentences which are formatted strings, in which the current keyword is placed. The layouts differ for some groups of keywords, as they can not all be placed within the same context logically.