Updated keyword identifier function
The previous keyword identifier function had some problems when the keyword was not identified. The function was divided into 3 parts. 1st identifies the keyword from the vector. 2nd runs it on the list of vectors and the last, main function executes them. The function returns true if the keyword is not identified.